# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Butter & Sugar",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "butter-sugar.zeabur.app",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "請填寫 Bearer token，例如: Bearer {token}"
    }
  },
  "paths": {
...
    "/api/v1/courses": {
      "get": {
        "tags": ["Courses"],
        "summary": "取得課程(測試用)",
        "description": "取得所有課程",
        "responses": {
          "200": {
            "description": "取得課程列表成功",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Courses"],
        "summary": "新增課程(測試用)",
        "description": "新增一筆課程資料",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "課程資料",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "teacher_id": {
                  "type": "string",
                  "example": "d02366a9-0c2d-480a-9685-bce659a2e079"
                },
                "category_id": {
                  "type": "string",
                  "example": "0e07a888-40a4-4338-9aed-4c38d6921ceb"
                },
                "course_name": {
                  "type": "string",
                  "example": "健康滿分！穀物燕麥餅乾入門與進階技巧"
                },
                "course_banner_imageUrl": {
                  "type": "string",
                  "example": "https://cdn.butterandsugar.com/courses/img2.jpg"
                },
                "course_banner_description": {
                  "type": "string",
                  "example": "一次學會低糖、無添加的健康燕麥餅乾，從基礎營養配方到創意穀物變化，讓美味和健康兼得！"
                },
                "course_description": {
                  "type": "string",
                  "example": "本課程專為注重健康的你設計，完整教學燕麥餅乾、堅果能量棒、全麥果乾餅乾等多種穀物點心。包含食材選擇、減糖配比、烘烤溫控及口感優化技巧，讓你做出低負擔、營養豐富又美味的家庭手作餅乾。"
                },
                "course_description_image": {
                  "type": "array",
                  "example": []
                },
                "course_smallimage": {
                  "type": "string",
                  "example": "https://cdn.butterandsugar.com/courses/img2.jpg"
                },
                "course_hours": {
                  "type": "integer",
                  "example": 8
                },
                "total_users": {
                  "type": "integer",
                  "example": 740
                },
                "trailer_vimeo_id": {
                  "type": "string",
                  "example": "butterSugar-00000002"
                },
                "trailer_name": {
                  "type": "string",
                  "example": "課程導覽"
                },
                "trailer_url": {
                  "type": "string",
                  "example": "https://vimeo.com/xyz9876abdc3.mp4"
                },
                "trailer_status": {
                  "type": "string",
                  "example": "ready"
                },
                "handout_name": {
                  "type": "string",
                  "example": "穀物燕麥餅乾講義"
                },
                "handout_url": {
                  "type": "string",
                  "example": "https://butterandsugar.com/handouts/handout2.pdf"
                },
                "suitable_for": {
                  "type": "string",
                  "example": "烘焙新手、健康飲食愛好者"
                },
                "course_goal": {
                  "type": "string",
                  "example": "學會健康燕麥餅乾的實用配方與多樣變化，輕鬆做出營養均衡的點心分享全家"
                },
                "origin_price": {
                  "type": "integer",
                  "example": 1000
                },
                "sell_price": {
                  "type": "integer",
                  "example": 850
                },
                "course_status": {
                  "type": "string",
                  "example": "上架"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "新增課程成功",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "新增課程成功"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "teacher_id": {
                      "type": "string",
                      "example": "d02366a9-0c2d-480a-9685-bce659a2e079"
                    },
                    "category_id": {
                      "type": "string",
                      "example": "0e07a888-40a4-4338-9aed-4c38d6921ceb"
                    },
                    "course_banner_imageUrl": {
                      "type": "string",
                      "example": "https://cdn.butterandsugar.com/courses/img2.jpg"
                    },
                    "course_name": {
                      "type": "string",
                      "example": "健康滿分！穀物燕麥餅乾入門與進階技巧"
                    },
                    "course_banner_description": {
                      "type": "string",
                      "example": "一次學會低糖、無添加的健康燕麥餅乾，從基礎營養配方到創意穀物變化，讓美味和健康兼得！"
                    },
                    "course_description": {
                      "type": "string",
                      "example": "本課程專為注重健康的你設計，完整教學燕麥餅乾、堅果能量棒、全麥果乾餅乾等多種穀物點心。包含食材選擇、減糖配比、烘烤溫控及口感優化技巧，讓你做出低負擔、營養豐富又美味的家庭手作餅乾。"
                    },
                    "course_description_image": {
                      "type": "array",
                      "example": []
                    },
                    "course_hours": {
                      "type": "integer",
                      "example": 8
                    },
                    "course_smallimage": {
                      "type": "string",
                      "example": "https://cdn.butterandsugar.com/courses/img2.jpg"
                    },
                    "total_users": {
                      "type": "integer",
                      "example": 740
                    },
                    "trailer_vimeo_id": {
                      "type": "string",
                      "example": "butterSugar-00000002"
                    },
                    "trailer_name": {
                      "type": "string",
                      "example": "課程導覽"
                    },
                    "trailer_url": {
                      "type": "string",
                      "example": "https://vimeo.com/xyz9876abdc3.mp4"
                    },
                    "trailer_status": {
                      "type": "string",
                      "example": "ready"
                    },
                    "handout_name": {
                      "type": "string",
                      "example": "穀物燕麥餅乾講義"
                    },
                    "handout_url": {
                      "type": "string",
                      "example": "https://butterandsugar.com/handouts/handout2.pdf"
                    },
                    "suitable_for": {
                      "type": "string",
                      "example": "烘焙新手、健康飲食愛好者"
                    },
                    "course_goal": {
                      "type": "string",
                      "example": "學會健康燕麥餅乾的實用配方與多樣變化，輕鬆做出營養均衡的點心分享全家"
                    },
                    "origin_price": {
                      "type": "integer",
                      "example": 1000
                    },
                    "sell_price": {
                      "type": "integer",
                      "example": 850
                    },
                    "course_status": {
                      "type": "string",
                      "example": "上架"
                    },
                    "id": {
                      "type": "string",
                      "example": "88d36f8d-fcea-4723-ad3f-5f959cfacc1a"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-06-04T12:00:17.404Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-06-04T12:00:17.404Z"
                    }
                  }
                }
              }
            }
          }
        }
      }
      },
...
    "/api/v1/cart/": {
      "get": {
        "tags": ["Cart"],
        "summary": "取得購物車內容",
        "description": "取得購物車內容",
        "responses": {
          "200": {
            "description": "取得購物車成功",
            "schema": {
              "type": "object",
              "properties": {
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Cart"],
        "summary": "加入課程到購物車",
        "description": "將指定課程加入購物車",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "course_id": {
                  "type": "string",
                  "example": "1c733528-e70f-42f6-948a-ea21abe9e36d"
                }
              },
              "required": ["course_id"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功加入課程到購物車",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "成功加入課程到購物車"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "item_count": {
                      "type": "integer",
                      "example": 1
                    },
                    "total_price": {
                      "type": "integer",
                      "example": 1200
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}